requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
certifi==2017.7.27.1
chardet==3.0.4
Click==7.0
cycler==0.10.0
dj-database-url==0.4.2
Django==1.11.6
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-geojson==2.12.0
django-leaflet==0.24.0
django-stringfield==0.3.0
django-timezone-field==3.0
flake8==3.5.0
gunicorn==19.7.1
idna==2.6
image==1.5.27
jsonfield==2.0.2
kiwisolver==1.0.1
kombu==4.5.0
matplotlib==3.0.2
mccabe==0.6.1
migrations==0.1.1
numpy==1.13.3
pandas==0.21.0
pew==1.1.0
Pillow==5.4.1
pipenv==8.3.1
psutil==5.3.1
psycopg2==2.7.3.2
PuLP==1.6.9
pvlib==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pyparsing==2.3.0
pysqlite3==0.2.1
python-crontab==2.3.6
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.2
redis==3.2.1
requests==2.18.4
rq==1.0
scipy==1.2.1
six==1.11.0
urllib3==1.22
vine==1.3.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
whitenoise==3.3.1
